Fixed major bugs in user interface and functions ...
- On function fromNFA in grammar class the algorithm was changed to
check for useless productions that were previously beeing made. Also
was added a check to see if the initial symbol is in any of the
productions right side. This change improves grammar readability because
it doesn't always create a new state S', this state is only created when
it is absolutely needed.
- Interface: Changes made to interface were ...
- Fixed method that loads the input grammar. Previously was removing \n
chars in the input string causing errors.
- NFA class: Now the operations functions create copy nfas of its
parameters and uses them in function keeping the parameters intact.
This was made because passing the same automata twice in the function
was causing erros.